worker/requirements.txt
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
